requirements.txt
# Requirements para Portfólio de Ciência de Dados
# Pós-Graduação UTFPR

# Análise de Dados
pandas>=1.5.0
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
imbalanced-learn>=0.10.0

# Visualização
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# Banco de Dados
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Geoespacial
geopandas>=0.13.0
folium>=0.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.20.0
notebook>=6.5.0

# Estatística
scipy>=1.10.0
statsmodels>=0.14.0

# Web Scraping e APIs
google-api-python-client>=2.0.0

# Processamento de Texto
nltk>=3.8.0
mlxtend>=0.22.0

# Utilitários
tqdm>=4.65.0
python-dotenv>=1.0.0

# LaTeX (opcional, para compilar artigos)
# Instalar separadamente: sudo apt-get install texlive-latex-extra texlive-fonts-recommended texlive-lang-portuguese

